Split PostgreSQL config to separate file
