+++ samples/server/go-server/api/server/storage/pgsqlconfig.go
@@ -1,0 +1,25 @@
+// Copyright 2017 The Grafeas Authors. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package storage
+
+type PgSQLConfig struct {
+	Host     string `yaml:"host"`
+	DbName   string `yaml:"dbname"`
+	User     string `yaml:"user"`
+	Password string `yaml:"password"`
+	// Valid sslmodes: disable, allow, prefer, require, verify-ca, verify-full.
+	// See https://www.postgresql.org/docs/current/static/libpq-connect.html for details
+	SSLMode string `yaml:"sslmode"`
+}
